refactor: 코인 추천 화면 객체 분리
